# Optimise JSON parsing