Some profiles at: http://dg.cx/geras-profile/. They are doing a query for 6 weeks of data at 5s resolution with two timeseries.

In particular this makes parsing go from ~1.1s to 360ms. Because it reduces the need for sorting it also makes savings in the main goroutine. Overall CPU time for me goes from 1.6s or so to ~400ms.

A little API break in that we don't copy the metric and tag names to every DataPoint (Geras doesn't care, also not sure why you'd want to do that)... The main thing left is the allocations for the many small DataPoint structures. Maybe could use sync.Pool or a better structure, but also this is probably enough for now.