@dgl 2019-10-26
It looks like the serialiser on the other end is custom and likely to always output sorted points: https://github.com/OpenTSDB/opentsdb/blob/b5133bd365a479cbc22f7f0cda888275575a04b6/src/tsd/HttpJsonSerializer.java#L847